requirements.txt
Django==2.2.6
django-filter==2.2.0
django-mysql==3.2.0
djangorestframework==3.10.3
Markdown==3.1.1
mysqlclient==1.3.13
pytz==2019.3
sqlparse==0.3.0
